+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+This project uses [Semantic Versioning 2.0.0](http://semver.org/).
+
 ## main
 
 ### Added